ANDROID: Android.bp: Build for /vendor
This often fits better as a vendor customization than a system tool.
When doing so, Soong complains:

  module trace-cmd missing dependencies: libdl{...} libm{...}, libc{...}

As a static executable, we'd need to statically link these libraries to
resolve the complaint, which seems excessive.

Instead, split out a "_static" variant as has been done in some other
packages, and make the main "trace-cmd" a dynamic ELF like most
binaries.

Bug: 393387208
Test: PRODUCT_PACKAGES_DEBUG += trace-cmd, build
Change-Id: Ifaaccfb1428c0f1532df2e6cbe4e72856cd51d31